arm: mxs: Add debug outputs and comments to mxs SPL source files
It is difficult to track down fail to boot issues in the mxs SPL.
Implement the following to make it easier:
 - Add debug outputs to allow tracing of SPL progress in order to track
where failure to boot occurs. DEUBUG and CONFIG_SPL_SERIAL_SUPPORT must
be defined to enable debug output in SPL
 - Add TODO comments where it is not clear if the code is doing what it
is meant to be doing, even tough the board boots properly (these comments
refer to existing code, not to any code added by this patch)

Signed-off-by: Graeme Russ <gr@denx.de>
---
Changes in v2
 - Add missing newlines
 - Add commit message

Changed in v3
 - None

Changes in v4
 - Changed Signed-off-by email address

Cover-letter:
Add support for booting mx28 boards without a battery
This series adds support for booting mx28 based boards which do not
include a battery as per Freescale application note AN4199

Patch 1 adds SPL debug output to help track down where early init
of the power block and SDRAM fails (define DEBUG and
CONFIG_SPL_SERIAL_SUPPORT in order to enable)

Patch 2 (which implements booting without a battery) is based on a patch
submitted to the Freescale community forums by Damien Gotfroi (define
CONFIG_SYS_MXS_VDD5V_ONLY to enable)

Patch 3 adds a useful halt upon completion of SPL in the case that the
board is booted in JTAG mode. If SPL debug is enabled, 'Waiting for JTAG
user' will be printed to the console when SPL has completed

Changes in v2
 - Dropped patch which adds Reachtech G2C1 board in order to allow the
   'no battery' functionality to be mainlined as soon as possible
 - Removed the patch which moved the PLL power up from spl_power_init.c
   to spl_mem_init.c - This patch will be considered in future rework
   of the power block and SDRAM initialisation code

Changes in v3
 - Clarified commit message for patch 2
 - Used explicit define in patch 2 (define == 0x00, so no functional
   change)
 - Removed extra blank line from patch #3

Changes in v4
 - Changed Signed-off-by email address
END

Series-cc: Damien Gotfroi <dgotfroi@greenwatch.be>

Series-prefix: v3